ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg[binary]==3.2.13
psycopg-pool==3.2.2
python-dotenv==1.0.0
sentence-transformers==3.0.0
scikit-learn==1.4.2
numpy==1.26.4
torch==2.2.0
pgvector==0.2.4
pydantic==2.5.1
pydantic-settings==2.1.0
requests==2.31.0
networkx==3.3
pytest==7.4.3
httpx==0.25.1
groq==1.2.0
